Halt parsnip and melon attack after picking plant
To make riskier plays easier, picking up melon or parsnip deletes their attack. This also makes it look less jarring as the attacks appear to be physically bound to the plant itself.
Plant is also now removed from the tree when picked, which removes it from the field while still allowing other nodes to reference it.